% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adm1_ne_nga}
\alias{adm1_ne_nga}
\title{administrative 1 (state) boundaries fro 3 states in NE Nigeria -}
\format{
A data frame with 3 rows and 3 variables:
\describe{
\item{ADM1_EN}{Admin 1 English name}
\item{ADM1_PCODE}{Admin 1 PCODE}
\item{geometry}{geometry}
...
}
}
\usage{
adm1_ne_nga
}
\description{
a polygon administrative 1 (state) boundaries fro 3 states in NE Nigeria -
}
\keyword{datasets}
