% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_extract_long.R
\name{ee_extract_long}
\alias{ee_extract_long}
\title{wrapper for \code{\link[rgee:ee_extract]{rgee::ee_extract()}} to help pivot from wide to long}
\usage{
ee_extract_long(ic, sf, sf_col, scale, reducer)
}
\arguments{
\item{ic}{ImageCollection}

\item{sf}{sf object for spatial disaggregation}

\item{sf_col}{sf column to include in output}

\item{scale}{scale (meters) of imageCollection}

\item{reducer}{sf_col level reducer}
}
\description{
wrapper for \code{\link[rgee:ee_extract]{rgee::ee_extract()}} to help pivot from wide to long
}
