% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_yr_mo_composite.R
\name{fun_sel}
\alias{fun_sel}
\title{function/reducer selecter - helper function}
\usage{
fun_sel(x)
}
\arguments{
\item{x}{reducer/statistic using typical r-syntax (character)}
}
\description{
function/reducer selecter - helper function
}
