% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_join_bands.R
\name{ic_join_bands}
\alias{ic_join_bands}
\title{ic_join_bands}
\usage{
ic_join_bands(x, y, by)
}
\arguments{
\item{x, y}{A pair of ImageCollections}

\item{by}{A character vector of variables to join by.}
}
\value{
An object of the same type as `x`. The output has the following properties:
Same number of images as `x`
Total number of bands equal the number of bands in `x` plus the number of bands in `y`
}
\description{
ic_join_bands
}
