% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_yr_mo_composite.R
\name{ic_yr_mo_composite_stat}
\alias{ic_yr_mo_composite_stat}
\title{make monthly composites from image collection}
\usage{
ic_yr_mo_composite_stat(
  ic,
  month_range = c(1, 12),
  year_range = c(2004, 2005),
  stat = "mean",
  monthly_stat_per = "year"
)
}
\arguments{
\item{ic}{ImageCollection}

\item{month_range}{months to create composites for}

\item{year_range}{years to consider}

\item{stat}{statistic/reducer to reduce ImageCollection by}

\item{monthly_stat_per}{time period to calculate monthly stat over. There are two options: `year` (default), or `range`.
if year is selected it will calculate each monthly statistic for every year set in `year_range`. If `range` is selected it will calculate 1 monthly statistic
across all years provided in `year_range`}
}
\description{
make monthly composites from image collection
}
