% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_date_to_band_name.R
\name{map_date_to_bandname_ic}
\alias{map_date_to_bandname_ic}
\title{map_date_to_bandname_ic}
\usage{
map_date_to_bandname_ic(ic)
}
\arguments{
\item{ic}{ImageCollection}
}
\description{
map_date_to_bandname_ic
}
