% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyrgee-package.R
\docType{package}
\name{easyrgee-package}
\alias{easyrgee}
\alias{easyrgee-package}
\title{easyrgee: easy rgee - wrappers for `rgee` functions}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
